Struct safe_nfs::directory_listing::DirectoryListing

pub struct DirectoryListing {
    // some fields omitted
}

DirectoryListing is the representation of a deserialised Directory in the network

Methods

impl DirectoryListing

fn new(name: String, tag_type: u64, user_metadata: Vec<u8>, versioned: bool, access_level: AccessLevel, parent_dir_key: Option<DirectoryKey>) -> Result<DirectoryListing, NfsError>

Create a new DirectoryListing

fn get_key(&self) -> &DirectoryKey

Returns the DirectoryKey representing the DirectoryListing

fn get_metadata(&self) -> &DirectoryMetadata

Get Directory metadata

fn get_mut_metadata(&mut self) -> &mut DirectoryMetadata

Get Directory metadata in mutable format so that it can also be updated

fn get_files(&self) -> &Vec<File>

Get all files in this DirectoryListing

fn get_mut_files(&mut self) -> &mut Vec<File>

Get all files in this DirectoryListing with mutability to update the listing of files

fn get_sub_directories(&self) -> &Vec<DirectoryMetadata>

Get all subdirectories in this DirectoryListing

fn get_mut_sub_directories(&mut self) -> &mut Vec<DirectoryMetadata>

Get all subdirectories in this DirectoryListing with mutability to update the listing of subdirectories

fn decrypt(client: Arc<Mutex<Client>>, directory_id: &XorName, data: Vec<u8>) -> Result<DirectoryListing, NfsError>

Decrypts a directory listing

fn encrypt(&self, client: Arc<Mutex<Client>>) -> Result<Vec<u8>, NfsError>

Encrypts the directory listing

fn find_file(&self, file_name: &String) -> Option<&File>

Get DirectoryInfo of sub_directory within a DirectoryListing. Returns the Option for the directory_name from the DirectoryListing

fn find_file_by_id(&self, id: &XorName) -> Option<&File>

Get DirectoryInfo of sub_directory within a DirectoryListing. Returns the Option for the directory_name from the DirectoryListing

fn find_sub_directory(&self, directory_name: &String) -> Option<&DirectoryMetadata>

Get DirectoryInfo of sub_directory within a DirectoryListing. Returns the Option for the directory_name from the DirectoryListing

fn find_sub_directory_by_id(&self, id: &XorName) -> Option<&DirectoryMetadata>

Get DirectoryInfo of sub_directory within a DirectoryListing. Returns the Option for the directory_name from the DirectoryListing

fn upsert_file(&mut self, file: File)

If file is present in the DirectoryListing then replace it else insert it

fn upsert_sub_directory(&mut self, directory_metadata: DirectoryMetadata)

If DirectoryMetadata is present in the sub_directories of DirectoryListing then replace it else insert it

fn remove_sub_directory(&mut self, directory_name: &String) -> Result<(), NfsError>

Remove a sub_directory

fn remove_file(&mut self, file_name: &String) -> Result<(), NfsError>

Remove a file

fn generate_nonce(directory_id: &XorName) -> Nonce

Generates a nonce based on the directory_id