Enum safe_nfs::errors::NfsError

pub enum NfsError {
    CoreError(CoreError),
    DirectoryAlreadyExistsWithSameName,
    DestinationAndSourceAreSame,
    DirectoryNotFound,
    FileAlreadyExistsWithSameName,
    FileDoesNotMatch,
    FileNotFound,
    InvalidRangeSpecified,
    ParameterIsNotValid,
    Unexpected(String),
    UnsuccessfulEncodeDecode(SerialisationError),
}

NFS Errors

Variants

CoreError

Client Error

DirectoryAlreadyExistsWithSameName

If Directory already exists with the same name in the same level

DestinationAndSourceAreSame

Destination is Same as the Source

DirectoryNotFound

Directory not found

FileAlreadyExistsWithSameName

File Already exists with the same name in a directory

FileDoesNotMatch

File does not match with the existing file in the directory listing

FileNotFound

File not found

InvalidRangeSpecified

Invalid byte range specified

ParameterIsNotValid

Validation error - if the field passed as parameter is not valid

Unexpected

Unexpected error

UnsuccessfulEncodeDecode

Unsuccessful Serialisation or Deserialisation