Struct safe_nfs::helper::directory_helper::DirectoryHelper

pub struct DirectoryHelper {
    // some fields omitted
}

DirectoryHelper provides helper functions to perform Operations on Directory

Methods

impl DirectoryHelper

fn new(client: Arc<Mutex<Client>>) -> DirectoryHelper

Create a new DirectoryHelper instance

fn create(&self, directory_name: String, tag_type: u64, user_metadata: Vec<u8>, versioned: bool, access_level: AccessLevel, parent_directory: Option<&mut DirectoryListing>) -> Result<(DirectoryListing, Option<DirectoryListing>), NfsError>

Creates a Directory in the network. When a directory is created and parent_directory is passed as a parameter. Then the parent directory is updated. If the parent_directory passed has a parent, then the parent_directory's parent is also updated and the same is returned Returns (created_directory, Option)

fn delete(&self, parent_directory: &mut DirectoryListing, directory_to_delete: &String) -> Result<Option<DirectoryListing>, NfsError>

Deletes a sub directory The parent_directory's parent is also updated if present Returns Option

fn update(&self, directory: &DirectoryListing) -> Result<Option<DirectoryListing>, NfsError>

Updates an existing DirectoryListing in the network. The parent_directory's parent is also updated and the same is returned Returns Option

fn get_versions(&self, directory_id: &NameType, type_tag: u64) -> Result<Vec<NameType>, NfsError>

Return the versions of the directory

fn get_by_version(&self, directory_id: &NameType, access_level: &AccessLevel, version: NameType) -> Result<DirectoryListing, NfsError>

Return the DirectoryListing for the specified version

fn get(&self, directory_key: &DirectoryKey) -> Result<DirectoryListing, NfsError>

Return the DirectoryListing for the latest version

fn get_user_root_directory_listing(&self) -> Result<DirectoryListing, NfsError>

Returns the Root Directory

fn get_configuration_directory_listing(&self, directory_name: String) -> Result<DirectoryListing, NfsError>

Returns the Configuration DirectoryListing from the configuration root folder Creates the directory or the root or both if it doesn't find one.