Struct safe_nfs::helper::file_helper::FileHelper

pub struct FileHelper {
    // some fields omitted
}

File provides helper functions to perform Operations on Files

Methods

impl FileHelper

fn new(client: Arc<Mutex<Client>>) -> FileHelper

Create a new FileHelper instance

fn create(&self, name: String, user_metatdata: Vec<u8>, parent_directory: DirectoryListing) -> Result<Writer, NfsError>

Helper function to create a file in a directory listing A writer object is returned, through which the data for the file can be written to the network The file is actually saved in the directory listing only after writer.close() is invoked

fn delete(&self, file_name: String, parent_directory: &mut DirectoryListing) -> Result<Option<DirectoryListing>, NfsError>

Delete a file from the DirectoryListing Returns Option

fn update_metadata(&self, file: File, parent_directory: &mut DirectoryListing) -> Result<Option<DirectoryListing>, NfsError>

Updates the file metadata. Returns Option

fn update_content(&self, file: File, mode: Mode, parent_directory: DirectoryListing) -> Result<Writer, NfsError>

Helper function to Update content of a file in a directory listing A writer object is returned, through which the data for the file can be written to the network The file is actually saved in the directory listing only after writer.close() is invoked

fn get_versions(&self, file: &File, parent_directory: &DirectoryListing) -> Result<Vec<File>, NfsError>

Return the versions of a directory containing modified versions of a file

fn read<'a>(&self, file: &'a File) -> Reader<'a>

Returns a reader for reading the file contents