Struct safe_nfs::helper::writer::Writer

pub struct Writer {
    // some fields omitted
}

Writer is used to write contents to a File and especially in chunks if the file happens to be too large

Methods

impl Writer

fn new(client: Arc<Mutex<Client>>, mode: Mode, parent_directory: DirectoryListing, file: File) -> Writer

Create new instance of Writer

fn write(&mut self, data: &[u8], position: u64)

Data of a file/blob can be written in smaller chunks

fn close(self) -> Result<(DirectoryListing, Option<DirectoryListing>), NfsError>

close is invoked only after all the data is completely written The file/blob is saved only when the close is invoked. Returns the update DirectoryListing which owns the file and also the updated DirectoryListing of the file's parent Returns (files's parent_directory, Option)