Struct safe_nfs::rest::Blob

pub struct Blob {
    // some fields omitted
}

Blob represents a File - Music, Video, Text etc

Methods

impl Blob

fn get_name(&self) -> &String

Get the name of the Blob

fn get_metadata(&self) -> String

Get the user settable Metadata of the Blob

fn get_created_time(&self) -> &Tm

Get the creation time for Blob

fn get_modified_time(&self) -> &Tm

Get the last modified time for the Blob

fn get_size(&self) -> u64

Get the Blob size in bytes

fn into_file(&self) -> &File

Convert the Blob to the format acceptable to the lower level Api's

fn into_mut_file(&mut self) -> &mut File

Convert the Blob to the format acceptable to the lower level Api's This can also be modified on the fly as the return is a mutable value