Struct safe_nfs::rest::Container

pub struct Container {
    // some fields omitted
}

Container Repersents a Directory. Container can have its own metadata, sub-containers and files

Methods

impl Container

fn authorise(client: Arc<Mutex<Client>>, container_info: Option<ContainerInfo>) -> Result<Container, NfsError>

Authorises the directory access. This sevrves as the initial access point of the Rest API. Operations can only be performed on a Container object. If the ContainerInfo parameter is None, then the user's root directory is returned. Returns the Container, if authorisation is successful.

fn create(&mut self, name: String, versioned: bool, access_level: AccessLevel, metadata: Option<String>) -> Result<(Container, Option<Container>), NfsError>

This functions is incoked to create a new container Say there are nested containers, Home - Pictures In the above example, Home is top level container and Pictures container is a child of Home When a new Conatiner Tour is created within Pictures, the following updates are carried out. 1. A new container is created 2. The metadata of the new Container (Tour) is added to the list of containers of Pictures. Modified time of Tour Container is also updated. 3. Metadata of Tour is updated in Home. Thus when a Container is created, the function returns the created Container and also the parent of the Container in which the new Container is being returned. Based on the above example, when the Container Tour is created in Pictures, this function would return a tpule of (Tour, Home) In case if there is no parent for the Container then None is returned. Returns tuple of created_container & parent_container of the the current

fn get_created_time(&self) -> &Tm

Returns the created time of the container

fn get_modified_time(&self) -> &Tm

Returns the last modified time of the container

fn get_info(&self) -> ContainerInfo

Return the unique id of the container

fn get_metadata(&self) -> String

Returns the user metadata saved as String.

fn get_name(&self) -> &String

Returns the name of the container

fn get_blobs(&self) -> Vec<Blob>

Returns the list of Blobs in the container

fn get_blob(&self, name: String) -> Result<Blob, NfsError>

Returns a Blob from the container

fn get_containers(&self) -> Vec<ContainerInfo>

Returns the list of child containers

fn update_metadata(&mut self, metadata: Option<String>) -> Result<Option<Container>, NfsError>

Updates the metadata of the container

fn get_versions(&self) -> Result<Vec<[u8; 64]>, NfsError>

Retrieves Versions for the container

fn get_container_versions(&self, container_info: &ContainerInfo) -> Result<Vec<[u8; 64]>, NfsError>

Retrieves Versions for the container being referred by the container_id

fn get_container(&mut self, container_info: &ContainerInfo, version: Option<[u8; 64]>) -> Result<Container, NfsError>

Fetches the latest version of the child container. Can fetch a specific version of the Container by passing the corresponding VersionId.

fn delete_container(&mut self, name: &String) -> Result<Option<Container>, NfsError>

Deletes the child container

fn create_blob(&mut self, name: String, metadata: Option<String>) -> Result<Writer, NfsError>

Creates a Blob within the container Returns a Writter object The content of the blob is written using the writter. The blob is created only after the writter.close() is invoked

fn update_blob_content(&mut self, blob: &Blob, data: &[u8]) -> Result<Option<Container>, NfsError>

Updates the blob content. Writes the complete data and updates the Blob

fn get_blob_writer(&mut self, blob: &Blob) -> Result<Writer, NfsError>

Return a writter object for the Blob, through which the content of the blob can be updated This is useful while handling larger files, to enable writting content in parts

fn get_blob_content(&self, blob: &Blob) -> Result<Vec<u8>, NfsError>

Reads the content of the blob and returns the complete content

fn get_blob_reader<'a>(&self, blob: &'a Blob) -> Result<Reader<'a>, NfsError>

Returns a reader for the blob Using a Reader helps in handling large file contents and also fetch data in a specific range

fn get_blob_versions(&self, name: &String) -> Result<Vec<Blob>, NfsError>

Returns the list of versions_id for the blob

fn update_blob_metadata(&mut self, blob: Blob, metadata: Option<String>) -> Result<Option<Container>, NfsError>

Update the metadata of the Blob in the container Returns Updated parent container, if the parent container exists.

fn delete_blob(&mut self, name: String) -> Result<(), NfsError>

Delete blob from the container

fn copy_blob(&mut self, blob_name: &String, to_container: &ContainerInfo) -> Result<(), NfsError>

Copies the latest blob version from the container to the specified destination container