Struct safe_nfs::helper::directory_helper::DirectoryHelper

pub struct DirectoryHelper {
    // some fields omitted
}

DirectoryHelper provides helper functions to perform Operations on Directory

Methods

impl DirectoryHelper

fn new(client: Arc<Mutex<Client>>) -> DirectoryHelper

Create a new DirectoryHelper instance

fn create(&self, directory_name: String, tag_type: u64, user_metadata: Vec<u8>, versioned: bool, access_level: AccessLevel, parent_directory: Option<&mut DirectoryListing>) -> Result<DirectoryListing, NfsError>

Creates a Directory in the network. Returns the created DirectoryListing

fn delete(&self, parent_directory: &mut DirectoryListing, directory_to_delete: &String) -> Result<(), NfsError>

Deletes a sub directory

fn update(&self, directory: &DirectoryListing) -> Result<DirectoryListing, NfsError>

Updates an existing DirectoryListing in the network. Returns the Updated DirectoryListing

fn update_directory_listing_and_parent(&self, directory: &DirectoryListing) -> Result<Option<DirectoryListing>, NfsError>

Updates an existing DirectoryListing in the network. Returns the Updated Parent DirectoryListing (if no parent then None is returned)

fn get_versions(&self, directory_key: (&NameType, u64)) -> Result<Vec<NameType>, NfsError>

Return the versions of the directory

fn get_by_version(&self, directory_key: (&NameType, u64), access_level: &AccessLevel, version: NameType) -> Result<DirectoryListing, NfsError>

Return the DirectoryListing for the specified version

fn get(&self, directory_key: (&NameType, u64), versioned: bool, access_level: &AccessLevel) -> Result<DirectoryListing, NfsError>

Return the DirectoryListing for the latest version

fn get_user_root_directory_listing(&self) -> Result<DirectoryListing, NfsError>

Returns the Root Directory

fn get_configuration_directory_listing(&self, directory_name: String) -> Result<DirectoryListing, NfsError>

Returns the Configuration DirectoryListing from the configuration root folder Creates the directory or the root or both if it doesn't find one.